Starting to rework mcts package.
It's in pieces at the moment, but things are converging.

Search algorithm modified to have only one epoch running on a graph at a
time.

Data model for search graph should now support up to (number of bits in
a usize) threads.

Still need to work out how this all comes together, particularly the
control of execution over multiple threads.